requirements.txt
python3-indy==1.8.0